third_party/blink/web_tests/external/wpt/fetch/private-network-access/window-open.tentative.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] private to local: failed preflight.
  assert_equals: expected "failure" but got "success"
[FAIL] private to local: missing CORS headers.
  assert_equals: expected "failure" but got "success"
[FAIL] private to local: missing PNA header.
  assert_equals: expected "failure" but got "success"
[FAIL] public to local: failed preflight.
  assert_equals: expected "failure" but got "success"
[FAIL] public to local: missing CORS headers.
  assert_equals: expected "failure" but got "success"
[FAIL] public to local: missing PNA header.
  assert_equals: expected "failure" but got "success"
[FAIL] public to private: failed preflight.
  assert_equals: expected "failure" but got "success"
[FAIL] public to private: missing CORS headers.
  assert_equals: expected "failure" but got "success"
[FAIL] public to private: missing PNA header.
  assert_equals: expected "failure" but got "success"
[FAIL] treat-as-public-address to local: failed preflight.
  assert_equals: expected "failure" but got "success"
[FAIL] treat-as-public-address to local: missing CORS headers.
  assert_equals: expected "failure" but got "success"
[FAIL] treat-as-public-address to local: missing PNA header.
  assert_equals: expected "failure" but got "success"
[FAIL] treat-as-public-address to private: failed preflight.
  assert_equals: expected "failure" but got "success"
[FAIL] treat-as-public-address to private: missing CORS headers.
  assert_equals: expected "failure" but got "success"
[FAIL] treat-as-public-address to private: missing PNA header.
  assert_equals: expected "failure" but got "success"
Harness: the test ran to completion.
